adjust synthea scripts to remove DB components, update bash script, move manual files to manual folder
remove quotes

Add end state data to manifest for pipeline validation
