+++ ops/ccs-ops-misc/synthetic-data/scripts/synthea-manual/generate-characteristics-file.py
