+++ ops/ccs-ops-misc/synthetic-data/scripts/synthea-manual/validate-synthea-load.py
